Implements ProcessMonitor in the haproxy driver
The ProcessMonitor class will monitor spawned external processes.

This patch enhances the HaproxyNSDriver class (v2) to utilize the
external_process module in order to monitor and respawn the haproxy processes
if and when needed.

With this patch the LBaaS agent (v2) will load external_process related options
in order to take a configured action when haproxy process dies unexpectedly.

This is the second attempt to introduce ProcessMonitor in the haproxy driver.
The first attempt I420ca20b2620487909885e0e9f08dae60ebec2bf caused an issue
where haproxy failed to acquire new configuration, due to the lack of -sf
flags which should be provided when haproxy.conf is being updated.
More information about the flags here[1]

[1] http://www.haproxy.org/download/1.2/doc/haproxy-en.txt

Co-Authored-By: Thomas Herve <therve@redhat.com>
Closes-Bug: #1565801
Related-Bug: #1603860

Change-Id: Ia47b1586be17be421e01c131859dd0d50b1d7db6